requirements.txt
aiosmtpd
alembic
aniso8601
atpublic
attrs
blinker
certifi
cffi
charset-normalizer
click
colorama
cryptography
dnspython
dominate
email-validator
Flask
Flask-Bootstrap
Flask-Login
Flask-Mail
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
Flask-WTF
greenlet
idna
itsdangerous
Jinja2
jsonschema
Mako
MarkupSafe
paho-mqtt
pycparser
PyJWT
pyrsistent
python-dotenv
pytz
requests
six
SQLAlchemy
urllib3
visitor
Werkzeug
WTForms
